// src/generated/t_tech/invest/grpc/marketdata.ts
export enum LastPriceType {
  /** LAST_PRICE_UNSPECIFIED - Не определен. */
  LAST_PRICE_UNSPECIFIED = 0,
  /** LAST_PRICE_EXCHANGE - Цена биржи. */
  LAST_PRICE_EXCHANGE = 1,
  /** LAST_PRICE_DEALER - Цена дилера */
  LAST_PRICE_DEALER = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface LastPrice {
  /** FIGI инструмента. */
  figi: string;
  /** Цена последней сделки за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  price:
    | Quotation
    | undefined;
  /** Время получения последней цены в часовом поясе UTC по времени биржи. */
  time:
    | Date
    | undefined;
  /** Тикер инструмента. */
  ticker: string;
  /** Класс-код (секция торгов). */
  classCode: string;
  /** UID инструмента. */
  instrumentUid: string;
  /** Тип последней цены. */
  lastPriceType: LastPriceType;
}
// ...
export const LastPrice: MessageFns<LastPrice> = {
  encode(message: LastPrice, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.figi !== "") {
      writer.uint32(10).string(message.figi);
    }
    if (message.price !== undefined) {
      Quotation.encode(message.price, writer.uint32(18).fork()).join();
    }
    if (message.time !== undefined) {
      Timestamp.encode(toTimestamp(message.time), writer.uint32(26).fork()).join();
    }
    if (message.ticker !== "") {
      writer.uint32(74).string(message.ticker);
    }
    if (message.classCode !== "") {
      writer.uint32(82).string(message.classCode);
    }
    if (message.instrumentUid !== "") {
      writer.uint32(90).string(message.instrumentUid);
    }
    if (message.lastPriceType !== 0) {
      writer.uint32(96).int32(message.lastPriceType);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LastPrice {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLastPrice();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.figi = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.price = Quotation.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.time = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.ticker = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.classCode = reader.string();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          message.instrumentUid = reader.string();
          continue;
        }
        case 12: {
          if (tag !== 96) {
            break;
          }

          message.lastPriceType = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LastPrice {
    return {
      figi: isSet(object.figi) ? globalThis.String(object.figi) : "",
      price: isSet(object.price) ? Quotation.fromJSON(object.price) : undefined,
      time: isSet(object.time) ? fromJsonTimestamp(object.time) : undefined,
      ticker: isSet(object.ticker) ? globalThis.String(object.ticker) : "",
      classCode: isSet(object.classCode)
        ? globalThis.String(object.classCode)
        : isSet(object.class_code)
        ? globalThis.String(object.class_code)
        : "",
      instrumentUid: isSet(object.instrumentUid)
        ? globalThis.String(object.instrumentUid)
        : isSet(object.instrument_uid)
        ? globalThis.String(object.instrument_uid)
        : "",
      lastPriceType: isSet(object.lastPriceType)
        ? lastPriceTypeFromJSON(object.lastPriceType)
        : isSet(object.last_price_type)
        ? lastPriceTypeFromJSON(object.last_price_type)
        : 0,
    };
  },

  toJSON(message: LastPrice): unknown {
    const obj: any = {};
    if (message.figi !== "") {
      obj.figi = message.figi;
    }
    if (message.price !== undefined) {
      obj.price = Quotation.toJSON(message.price);
    }
    if (message.time !== undefined) {
      obj.time = message.time.toISOString();
    }
    if (message.ticker !== "") {
      obj.ticker = message.ticker;
    }
    if (message.classCode !== "") {
      obj.classCode = message.classCode;
    }
    if (message.instrumentUid !== "") {
      obj.instrumentUid = message.instrumentUid;
    }
    if (message.lastPriceType !== 0) {
      obj.lastPriceType = lastPriceTypeToJSON(message.lastPriceType);
    }
    return obj;
  },

  create(base?: DeepPartial<LastPrice>): LastPrice {
    return LastPrice.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<LastPrice>): LastPrice {
    const message = createBaseLastPrice();
    message.figi = object.figi ?? "";
    message.price = (object.price !== undefined && object.price !== null)
      ? Quotation.fromPartial(object.price)
      : undefined;
    message.time = object.time ?? undefined;
    message.ticker = object.ticker ?? "";
    message.classCode = object.classCode ?? "";
    message.instrumentUid = object.instrumentUid ?? "";
    message.lastPriceType = object.lastPriceType ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}